Split dependencies file into smaller files with related logic
